do not downgrade when group contains non-finite lookahead
We should downgrade a reference if the group it points at is non-finite size. Previously if the group was finite size, but contained a lookahead that wasn't, we would consider the group non-finite. This is now fixed.

Also does some refactoring to remove `groups`, `lookaheadStack` and `quantifierStack`. `stack` contains everything that was in those in one array.