Always stop expanding quantifier after min iterations after loop detected
...unless the quantifier is part of a group that is referenced later.

Previously if there was a max then we would always generate up to the max iterations. If the max was infinity then we would stop once the sequence of the left against right hit an infinite loop.

Now we always stop when the sequence of left against right hits an infinite loop, once the min has been reached, unless it's part of a group referenced later.